requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
chardet==4.0.0
discord.py==1.6.0
future==0.18.2
idna==3.1
importlib-metadata==3.7.3
jedi==0.17.0
multidict==5.1.0
parso==0.8.1
pluggy==0.13.1
python-jsonrpc-server==0.4.0
python-language-server==0.35.1
typing-extensions==3.7.4.3
ujson==4.0.2
websockets==8.1
yarl==1.6.3
zipp==3.4.1
matplotlib
pandas
mplfinance
requests
